refactor: ApiClient의 모듈 분리
