+++ app/src/main/java/com/wasingun/seller_lounge/di/ApiClientModule.kt
@@ -1,0 +1,60 @@
+package com.wasingun.seller_lounge.di
+
+import com.google.gson.Gson
+import com.wasingun.seller_lounge.BuildConfig
+import com.wasingun.seller_lounge.data.source.NaverApiClient
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ApiClientModule {
+    private const val BASE_URL = "https://openapi.naver.com/v1/"
+    private val gson = Gson()
+
+    @Singleton
+    @Provides
+    fun provideOkHttpClient(): OkHttpClient {
+        val logger = HttpLoggingInterceptor().apply {
+            level = HttpLoggingInterceptor.Level.BODY
+        }
+
+        val header = Interceptor { chain ->
+            val newRequest = chain.request().newBuilder()
+                .addHeader("X-Naver-Client-Id", BuildConfig.NAVER_CLIENT_ID)
+                .addHeader("X-Naver-Client-Secret", BuildConfig.NAVER_CLIENT_SECRET)
+                .addHeader("Content-Type", "application/json")
+                .build()
+            chain.proceed(newRequest)
+        }
+
+        return OkHttpClient.Builder()
+            .addInterceptor(logger)
+            .addInterceptor(header)
+            .build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideRetrofit(okHttp: OkHttpClient): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .client(okHttp)
+            .addConverterFactory(GsonConverterFactory.create(gson))
+            .build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideApiClient(retrofit: Retrofit): NaverApiClient {
+        return retrofit.create(NaverApiClient::class.java)
+    }
+}
